pcs/services/device-telemetry/Services.Test/AlarmsTest.cs
// Copyright (c) Microsoft. All rights reserved.

using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.Azure.Documents;
using Microsoft.Azure.IoTSolutions.DeviceTelemetry.Services;
using Microsoft.Azure.IoTSolutions.DeviceTelemetry.Services.Diagnostics;
using Microsoft.Azure.IoTSolutions.DeviceTelemetry.Services.Exceptions;
using Microsoft.Azure.IoTSolutions.DeviceTelemetry.Services.Runtime;
using Microsoft.Azure.IoTSolutions.DeviceTelemetry.Services.Storage.CosmosDB;
using Moq;
using Services.Test.helpers;
using Xunit;

namespace Services.Test
{
    public class AlarmsTest
    {
        private readonly Mock<IStorageClient> storageClient;
        private readonly Mock<ILogger> logger;
        private readonly IAlarms alarms;

        public AlarmsTest()
        {
            var servicesConfig = new ServicesConfig
            {
                AlarmsConfig = new AlarmsConfig("database", "collection", 3)
            };
            this.storageClient = new Mock<IStorageClient>();
            this.logger = new Mock<ILogger>();
            this.alarms = new Alarms(servicesConfig, this.storageClient.Object, this.logger.Object);
        }

        /**
         * Test basic functionality of delete alarms by id.
         */
        [Fact, Trait(Constants.TYPE, Constants.UNIT_TEST)]
        public void BasicDelete()
        {
            // Arrange
            List<string> ids = new List<string> { "id1", "id2", "id3", "id4" };
            Document d1 = new Document
            {
                Id = "test"
            };
            this.storageClient
                .Setup(x => x.DeleteDocumentAsync(
                    It.IsAny<string>(),
                    It.IsAny<string>(),
                    It.IsAny<string>()))
                .Returns(Task.FromResult(d1));

            // Act
            this.alarms.Delete(ids);

            // Assert
            for (int i = 0; i < ids.Count; i++)
            {
                this.storageClient.Verify(x => x.DeleteDocumentAsync("database", "collection", ids[i]), Times.Once);
            }

            this.logger.Verify(l => l.Error(It.IsAny<string>(), It.IsAny<Func<object>>()), Times.Never);
            this.logger.Verify(l => l.Warn(It.IsAny<string>(), It.IsAny<Func<object>>()), Times.Never);
        }

        /**
         * Verify if delete alarm by id fails once it will retry
        */
        [Fact, Trait(Constants.TYPE, Constants.UNIT_TEST)]
        public async Task DeleteSucceedsTransientExceptionAsync()
        {
            // Arrange
            List<string> ids = new List<string> { "id1" };
            Document d1 = new Document
            {
                Id = "test"
            };
            this.storageClient
                .SetupSequence(x => x.DeleteDocumentAsync(
                    It.IsAny<string>(),
                    It.IsAny<string>(),
                    It.IsAny<string>()))
                .Throws(new Exception())
                .Returns(Task.FromResult(d1));

            // Act
            await this.alarms.Delete(ids);

            // Assert
            this.storageClient.Verify(x => x.DeleteDocumentAsync("database", "collection", ids[0]), Times.Exactly(2));

            this.logger.Verify(l => l.Error(It.IsAny<string>(), It.IsAny<Func<object>>()), Times.Never);
            this.logger.Verify(l => l.Warn(It.IsAny<string>(), It.IsAny<Func<object>>()), Times.Once);
        }

        /**
         * Verify that after 3 failures to delete an alarm an
         * exception will be thrown.
         */
        [Fact, Trait(Constants.TYPE, Constants.UNIT_TEST)]
        public async Task DeleteFailsAfter3ExceptionsAsync()
        {
            // Arrange
            List<string> ids = new List<string> { "id1" };
            Document d1 = new Document
            {
                Id = "test"
            };

            this.storageClient
                .SetupSequence(x => x.DeleteDocumentAsync(
                    It.IsAny<string>(),
                    It.IsAny<string>(),
                    It.IsAny<string>()))
                .Throws(new Exception())
                .Throws(new Exception())
                .Throws(new Exception());

            // Act
            await Assert.ThrowsAsync<ExternalDependencyException>(async () => await this.alarms.Delete(ids));

            // Assert
            this.storageClient.Verify(x => x.DeleteDocumentAsync("database", "collection", ids[0]), Times.Exactly(3));

            this.logger.Verify(l => l.Error(It.IsAny<string>(), It.IsAny<Func<object>>()), Times.Once);
            this.logger.Verify(l => l.Warn(It.IsAny<string>(), It.IsAny<Func<object>>()), Times.Exactly(2));
        }

        [Fact, Trait(Constants.TYPE, Constants.UNIT_TEST)]
        public async Task ThrowsOnInvalidInput()
        {
            // Arrange
            var xssString = "<body onload=alert('test1')>";
            var xssList = new List<string>
            {
                "<body onload=alert('test1')>",
                "<IMG SRC=j&#X41vascript:alert('test2')>"
            };

            // Act & Assert
            await Assert.ThrowsAsync<InvalidInputException>(async () => await this.alarms.DeleteAsync(xssString));
            await Assert.ThrowsAsync<InvalidInputException>(async () => await this.alarms.Delete(xssList));
            await Assert.ThrowsAsync<InvalidInputException>(async () => await this.alarms.UpdateAsync(xssString, xssString));
            Assert.Throws<InvalidInputException>(() => this.alarms.GetCountByRule(xssString, DateTimeOffset.MaxValue, DateTimeOffset.MaxValue, xssList.ToArray()));
            Assert.Throws<InvalidInputException>(() => this.alarms.List(null, null, xssString, 0, 1, xssList.ToArray()));
            Assert.Throws<InvalidInputException>(() => this.alarms.ListByRule(xssString, DateTimeOffset.MaxValue, DateTimeOffset.MaxValue, xssString, 0, 1, xssList.ToArray()));
        }
    }
}
